+++ mtg/src/test/scala/com/github/gchudnov/mtg/ordering/IntervalOrderingSpec.scala
@@ -1,7 +1,9 @@
-package com.github.gchudnov.mtg
+package com.github.gchudnov.mtg.ordering
 
 import com.github.gchudnov.mtg.Interval
 import com.github.gchudnov.mtg.Domain
+import com.github.gchudnov.mtg.Mark
+import com.github.gchudnov.mtg.Value
 import com.github.gchudnov.mtg.Arbitraries.*
 import com.github.gchudnov.mtg.TestSpec
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks.*
