+++ mtg/src/test/scala/com/github/gchudnov/mtg/ordering/MarkOrderingSpec.scala
@@ -1,6 +1,8 @@
-package com.github.gchudnov.mtg
+package com.github.gchudnov.mtg.ordering
 
 import com.github.gchudnov.mtg.Interval
+import com.github.gchudnov.mtg.Mark
+import com.github.gchudnov.mtg.Value
 import com.github.gchudnov.mtg.Domain
 import com.github.gchudnov.mtg.TestSpec
